chatai_api/pyproject.toml
[project]
name = "chatai-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.0.17",
    "faiss-cpu>=1.12.0",
    "fastapi>=0.116.1",
    "numpy>=2.3.2",
    "openai>=1.99.9",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "sentence-transformers>=5.1.0",
    "uvicorn>=0.35.0",
]
